+++ src/scrappers/html/index.js
@@ -1,0 +1,3 @@
+export { default as config } from "./config-puppeteer";
+export { fetchHtml, fetchHtmlWithCache, fetchHtmlWithPuppeteer } from "./fetch";
+export { htmlDOM } from "./domParser";
