+++ src/scrappers/source/proCyclingStats/index.js
@@ -1,4 +1,3 @@
-export { default as config } from "./config";
 export { interceptRequests } from "./page_proCyclingStats";
 export { collectWorldTourRaces, scrapeRaces } from "./races";
 export { scrapeRaceStartList } from "./raceStartList";
